Fixed bug in new Task creation.
The new Task creation form does not use exclamation marks to
represent priority but natural language entries from lang[..].
This caused ~~TASK?[priority]~~ entries in pages created by this
form to not conform to the syntax and both priority and due date
(if given in the form) were not correctly entered.

I added a "translation step" before new page creation that should
fix this (haven't tested for different languages, but it should
work).